backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
